requirements.txt


django==2.2.10
coreapi==2.3.3
pyyaml==5.1
django-rest-swagger==2.2.0
django-cors-headers==3.4.0
djangorestframework==3.11.1

